move user repository interface and implementation to own folder
